requirements.txt
json
google-cloud-bigquery
google-auth
google-auth-oauthlib